requirements.txt
streamlit
tensorflow
numpy==1.26.4
opencv-python
pillow
scikit-learn
matplotlib
